pyproject.toml
[tool.poetry]
name = "litellm-agents"
version = "0.1.0"
description = "Multi-agent system using LiteLLM with OpenAI and Ollama support"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
litellm = "^1.74.0.post1"
langgraph = "^0.5.2"
pydantic = "^2.11.7"
python-dotenv = "^1.1.1"
click = "^8.2.1"
instructor = "^1.9.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-asyncio = "^1.0.0"

